fix(dispatch): multiplier overflow 가드 + 타입 int로 좁히기 + 생성자 인자 검증 테스트 부활

+++ src/main/java/com/example/alarm/config/DispatchProperties.java
@@ -25,7 +25,7 @@ public class DispatchProperties {
     private long backoffBaseMs = 1_000;
     private long backoffMaxMs = 30L * 60 * 1000;
     private double backoffJitterRatio = 0.2;
-    private long backoffMultiplier = 2;
+    private int backoffMultiplier = 2;
     private int executorAwaitSeconds = 30;
 
     /**
